+++ crates/samples/reactor/swap_chain_panel/src/main.rs
@@ -26,7 +26,7 @@ fn create_d3d_swap_chain(
             HMODULE::default(),
             D3D11_CREATE_DEVICE_BGRA_SUPPORT as u32,
             Some(&[D3D_FEATURE_LEVEL_11_0]),
-            D3D11_SDK_VERSION,
+            D3D11_SDK_VERSION as u32,
             Some(&mut device),
             None,
             Some(&mut context),
