+++ crates/tests/libs/registry/tests/transaction.rs
@@ -25,7 +25,7 @@ fn create_with_transaction() {
     // The transaction is not yet committed so this non-transaction read will fail.
     assert_eq!(
         key.get_u64("u64").unwrap_err().code(),
-        WIN32_ERROR(ERROR_FILE_NOT_FOUND).into()
+        WIN32_ERROR(ERROR_FILE_NOT_FOUND as u32).into()
     );
 
     tx.commit().unwrap();
@@ -36,6 +36,6 @@ fn create_with_transaction() {
     // The transaction is no longer active so this key cannot be used.
     assert_eq!(
         tx_key.get_u64("u64").unwrap_err().code(),
-        WIN32_ERROR(ERROR_TRANSACTION_NOT_ACTIVE).into()
+        WIN32_ERROR(ERROR_TRANSACTION_NOT_ACTIVE as u32).into()
     );
 }
