+++ crates/tests/libs/strings/tests/pcstr.rs
@@ -13,7 +13,10 @@ fn test() -> Result<()> {
     let invalid = &[0xc0, 0x80];
     let p = PCSTR::from_raw(invalid.as_ptr());
     let e: Error = unsafe { p.to_string().unwrap_err().into() };
-    assert_eq!(e.code(), WIN32_ERROR(ERROR_NO_UNICODE_TRANSLATION).into());
+    assert_eq!(
+        e.code(),
+        WIN32_ERROR(ERROR_NO_UNICODE_TRANSLATION as u32).into()
+    );
     assert_eq!(
         e.message(),
         "No mapping for the Unicode character exists in the target multi-byte code page."
