+++ crates/tests/libs/strings/tests/pcwstr.rs
@@ -13,7 +13,10 @@ fn test() -> Result<()> {
     let invalid = &[0xD834, 0xDD1E, 0x006d, 0x0075, 0xD800, 0x0069, 0x0063];
     let p = PCWSTR::from_raw(invalid.as_ptr());
     let e: Error = unsafe { p.to_string().unwrap_err().into() };
-    assert_eq!(e.code(), WIN32_ERROR(ERROR_NO_UNICODE_TRANSLATION).into());
+    assert_eq!(
+        e.code(),
+        WIN32_ERROR(ERROR_NO_UNICODE_TRANSLATION as u32).into()
+    );
     assert_eq!(
         e.message(),
         "No mapping for the Unicode character exists in the target multi-byte code page."
